finished

+++ blockingqueue.h
@@ -11,10 +11,25 @@ typedef struct _data
 
 typedef struct bqueue_ *blockingqueue;
 
+/**
+ * @brief Initialize internal resources of the object
+ */
 blockingqueue bqueue_init(size_t content_size);
+/**
+ * @brief enqueue data into the queue
+ */
 void enqueue(blockingqueue, void *data_ptr);
+/**
+ * @brief dequeue data from the queue or blocks until available data
+ */
 void dequeue(blockingqueue, void *data_ptr);
-size_t get_size(blockingqueue);
+/**
+ * @brief Get the size of the queue
+ */
+int get_size(blockingqueue);
+/**
+ * @brief Release internal resources
+ */
 void bqueue_destroy(blockingqueue);
 
 #endif
